Introduce an overwatching application context
If there is active application context, we need to have an
application context that can monitor the execution of the tests.
The test application context is started late in the test execution
lifecycle and thus misses events in the initializing phase. This will be
used in the future to record the events and to generate reports and such