Adds persistence of test run events
The events are stored using a basic form of event sourcing. Each event
is converted to a "cloud event" (see https://cloudevents.io/), which is
a canonical form of event representation.

This also introduces an "app" submodule which does the event handling
and persisting. In the future this will be used to host the web
API and interface.