Launch the autotest Spring application if no context exists
Instead of launching nested Spring Boot contexts for each test and
instead of launching a dummy autotest application
(ApplicationContextLifecycleApplication) if none exist we now launch the
user's own application at the beginning of the test. We do this by
trying to figure out the main method using Spring Boot's own
MainMethodFinder on the class path and simply invoke the main method.